[asset-tracker] number format issue resolved, stock added, customTooltip
